backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
bcrypt==4.3.0
bidict==0.23.1
black==23.11.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
fastapi>=0.110.1
flake8==6.1.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
isort==5.12.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
orjson==3.9.10
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyflakes==3.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.18
python-socketio==5.14.0
PyYAML==6.0.2
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette>=0.36.3
structlog==23.2.0
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
aiosqlite==0.20.0
# Upload and OCR dependencies
boto3==1.35.0
celery==5.4.0
redis==5.1.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
